+++ packages/engine-core/src/index.ts
@@ -4,3 +4,5 @@ export * from "./elrc.js";
 export * from "./math.js";
 export * from "./types.js";
 export * from "./typographySelectors.js";
+export * from "./typographyComposition.js";
+export * from "./colorHarmony.js";